+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.
 * Admins can add users to groups even if users have pending invitations ([#773](https://gitlab.com/yale-sdmp/vesta/issues/773)).
 * Unhidden table caption is now hidden ([#800](https://gitlab.com/yale-sdmp/vesta/issues/800)).
 * Fix NoMethodError in creating new user in admin dashboard ([#810](https://gitlab.com/yale-sdmp/vesta/issues/810))
+* Properly handle a failed result in IDRProfileQuerier ([#689](https://gitlab.com/yale-sdmp/vesta/issues/689)).
 
 ## v1.0.8 - 2018-04-09
 ### Fixed